scan tool now skips over generated email files, only shows results for dirty word hits

+++ pages/views/scan.py
@@ -19,6 +19,9 @@ from pdfminer.high_level import *
 
 # cfts settings
 from cfts import settings as cftsSettings
+
+# regular expressions
+import re
 # ====================================================================
 
 
@@ -32,7 +35,6 @@ def scan(request,pullZip):
     # POST
     if request.method == 'POST':
         if pullZip !="none":
-            print("Scanning from pull")
             pullZip = os.path.join(cftsSettings.BASE_DIR,"pulls",pullZip)
             zf = ZipFile(pullZip)
             zf.extractall(settings.SCANTOOL_DIR)
@@ -48,17 +50,6 @@ def scan(request,pullZip):
         # clean up after yourself
         cleanup(settings.SCANTOOL_DIR)
         return JsonResponse(results, safe=False)
-    
-    # if request.method == 'GET':
-    #     if pullNet !="none" and pullNum !="none":
-    #         pullZip = os.path.join(cftsSettings.BASE_DIR,"pulls",pullNet+"_"+pullNum+".zip")
-    #         zf = ZipFile(pullZip)
-    #         zf.extractall(settings.SCANTOOL_DIR)
-    #         results = runScan()
-
-    #         # clean up after yourself
-    #         cleanup(settings.SCANTOOL_DIR)
-    #         return render(request, 'pages/scan.html', {'rc': rc, 'results': results})
 
     # GET
     if pullZip !="none":
@@ -75,37 +66,46 @@ def runScan():
 
     # \cfts\scan should contain all the user folders from the zip file
     for root, subdirs, files in os.walk(scan_dir):
+        txt = re.compile('_email(\d+)?.txt')
+        eml = re.compile('_email(\d+)?.eml')
+
         for filename in files:
-            file_results = None
-            file_path = os.path.join(root, filename)
-##      print( '\t- file %s (full path: %s)' % ( filename, file_path ) )
-            temp, ext = os.path.splitext(filename)
+            if txt.match(filename) == None and eml.match(filename) == None:
+                file_results = None
+                file_path = os.path.join(root, filename)
+    ##      print( '\t- file %s (full path: %s)' % ( filename, file_path ) )
+                temp, ext = os.path.splitext(filename)
 
-            if(ext in office_filetype_list):
-                file_results = scanOfficeFile(file_path)
-            elif(ext == '.pdf'):
-                textFile = open(root+"\\"+temp+".txt", "w")
-                pdf2Text = StringIO()
-                with open(file_path, 'rb') as pdf:
-                    extract_text_to_fp(
-                        pdf, pdf2Text, output_type='text', codec='utf-8')
-                    textFile.write(pdf2Text.getvalue().strip())
-                    textFile.close()
+                if(ext in office_filetype_list):
+                    file_results = scanOfficeFile(file_path)
+                elif(ext == '.pdf'):
+                    textFile = open(root+"\\"+temp+".txt", "w")
+                    pdf2Text = StringIO()
+                    with open(file_path, 'rb') as pdf:
+                        extract_text_to_fp(
+                            pdf, pdf2Text, output_type='text', codec='utf-8')
+                        textFile.write(pdf2Text.getvalue().strip())
+                        textFile.close()
 
-                text_path = os.path.join(root+"\\"+temp+".txt")
-                file_results = scanFile(text_path)
-                file_results['file'] = file_path
-                os.remove(text_path)
+                    text_path = os.path.join(root+"\\"+temp+".txt")
+                    file_results = scanFile(text_path)
+                    if file_results is not None:
+                        file_results['file'] = file_path
+                    os.remove(text_path)
 
+                else:
+                    file_results = scanFile(file_path)
+
+                if(file_results is not None):
+                    result = {}
+                    result['file'] = file_path
+                    result['found'] = file_results
+                    scan_results.append(result)
+                else:
+                    print("no hits for file: ", filename)
             else:
-                file_results = scanFile(file_path)
-
-            if(file_results is not None):
-                result = {}
-                result['file'] = file_path
-                result['found'] = file_results
-                scan_results.append(result)
-
+                print("Skipping email file...:", filename)
+    print("Scan results: ", scan_results)
     return scan_results
 
 
@@ -132,10 +132,11 @@ def scanOfficeFile(office_file):
 
 
 def scanFile(text_file):
-    result = {
-        'file': text_file,
-        'findings': []
-    }
+    # result = {
+    #     'file': text_file,
+    #     'findings': []
+    # }
+    result = None
 
     dirty_word_list = ["SECRET", "S//", "NOFORN",
                        "C//", "CONFIDENTIAL", "PRV", "LVY"]
@@ -146,10 +147,10 @@ def scanFile(text_file):
     try:
         with open(text_file, "r", encoding="utf-8") as f:
             f_content = f.read()
-            # result = {
-            #    'file': text_file,
-            #    'findings': []
-            # }
+            result = {
+               'file': text_file,
+               'findings': []
+            }
             for compiled_reg in reg_lst:
                 found = re.finditer(compiled_reg, f_content)
                 for match in found:
@@ -157,7 +158,7 @@ def scanFile(text_file):
                     result['findings'].append(
                         'This file contains the term: %s' % (match.group()))
             if not len(result['findings']):
-                result['findings'].append("Scan returned no dirty words.")
+                result = None
 
     except UnicodeDecodeError:
         # Found non-text data
